requirements.txt
fastapi
pyjwt
python-dotenv
pydantic
sqlalchemy
bcrypt
pymysql
pydantic
